Fold payments module into games app instead of a separate app
Moves settlement/payment models, views, helpers, mailer, admin, and
templates from the standalone payments app into games, since the
feature is small and tightly coupled to games/players. Drops the
payments app registration and its URL include.
