+++ futsal_app/settings.py
@@ -38,7 +38,6 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "games",
-    "payments",
 ]
 
 MIDDLEWARE = [
