+++ futsal_app/urls.py
@@ -23,6 +23,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("", RedirectView.as_view(url="games/")),
     path("games/", include("games.urls")),
-    path("payments/", include("payments.urls")),
     path("accounts/", include("django.contrib.auth.urls")),
 ]
