+++ games/admin.py
@@ -1,6 +1,13 @@
 from django.contrib import admin
 
-from .models import BookingHistoryForGame, Game, Player
+from .models import (
+    BookingHistoryForGame,
+    Game,
+    GamePrice,
+    Player,
+    PlayerCharge,
+    SettlementRun,
+)
 
 
 class GameAdmin(admin.ModelAdmin):
@@ -9,7 +16,28 @@ class GameAdmin(admin.ModelAdmin):
     date_hierarchy = "when"
 
 
+class SettlementRunAdmin(admin.ModelAdmin):
+    list_display = ("year", "month", "send_count", "last_sent_at", "created_at")
+    list_filter = ("year", "month")
+
+
+class PlayerChargeAdmin(admin.ModelAdmin):
+    list_display = (
+        "settlement_run",
+        "player",
+        "amount",
+        "game_count",
+        "is_paid",
+        "paid_at",
+        "marked_by",
+    )
+    list_filter = ("is_paid", "settlement_run")
+
+
 # Register your models here.
 admin.site.register(Game, GameAdmin)
 admin.site.register(BookingHistoryForGame)
 admin.site.register(Player)
+admin.site.register(GamePrice)
+admin.site.register(SettlementRun, SettlementRunAdmin)
+admin.site.register(PlayerCharge, PlayerChargeAdmin)
