+++ games/mailer.py
@@ -4,7 +4,7 @@ from django.template.loader import render_to_string
 
 from futsal_app import settings
 from games.helpers import game_helper, player_helper
-from games.models import Game, Player
+from games.models import Game, Player, PlayerCharge
 
 User = get_user_model()
 
@@ -144,3 +144,29 @@ def send_player_status_update_email_to_admins(player: Player, game: Game, status
         msg = EmailMultiAlternatives(subject, text_content, from_email, to)
         msg.attach_alternative(html_content, "text/html")
         msg.send()
+
+
+def send_settlement_email(charge: PlayerCharge) -> None:
+    run = charge.settlement_run
+    player_display_name = player_helper.get_display_name(charge.player)
+    subject = f"Rozliczenie za {run.month:02d}/{run.year}"
+    from_email = settings.DEFAULT_FROM_EMAIL
+    to = [charge.player.user.email]
+
+    text_content = (
+        f"Hello {player_display_name}, your settlement for {run.month:02d}/{run.year} "
+        f"is {charge.amount} for {charge.game_count} game(s)."
+    )
+
+    html_content = render_to_string(
+        "emails/settlement_email.html",
+        {
+            "player_display_name": player_display_name,
+            "charge": charge,
+            "run": run,
+        },
+    )
+
+    msg = EmailMultiAlternatives(subject, text_content, from_email, to)
+    msg.attach_alternative(html_content, "text/html")
+    msg.send()
