+++ games/urls.py
@@ -24,6 +24,14 @@ urlpatterns = [
     path("add_game", views.add_game, name="add_game_url"),
     path("add_absence", views.add_absence, name="add_absence_url"),
     path("booking_history/", views.booking_history, name="booking_history_url"),
+    path("payments/", views.settlement_overview, name="settlement_overview_url"),
+    path("payments/send/", views.send_settlement, name="send_settlement_url"),
+    path(
+        "payments/charge/<int:charge_id>/toggle-paid/",
+        views.toggle_paid,
+        name="toggle_paid_url",
+    ),
+    path("payments/who-paid/", views.who_paid, name="who_paid_url"),
     path("logout/", LogoutView.as_view(next_page="/accounts/login"), name="logout_url"),
     path("login/", LoginView.as_view(next_page="next_games_url"), name="login_url"),
     path(
